Extract legacy map codec ownership
